+++ code/package.json
@@ -29,7 +29,7 @@
     "check": "NODE_ENV=production node ../scripts/check-package.js",
     "ci-tests": "yarn task --task check --no-link --start-from=install && yarn lint && yarn test && cd ../scripts && yarn test",
     "danger": "danger",
-    "generate-sandboxes": "ts-node --swc ../scripts/sandbox-generators/generate.ts",
+    "generate-sandboxes": "ts-node --swc ../scripts/sandbox/generate.ts",
     "github-release": "github-release-from-changelog",
     "linear-export": "ts-node --swc --project=../scripts/tsconfig.json ../scripts/linear-export.ts",
     "lint": "yarn lint:js && yarn lint:md",
@@ -40,7 +40,7 @@
     "lint:other": "prettier --write '**/*.{css,html,json,md,yml}'",
     "lint:package": "sort-package-json",
     "local-registry": "ts-node --swc --project=../scripts/tsconfig.json ../scripts/run-registry.ts",
-    "publish-sandboxes": "ts-node --swc ../scripts/sandbox-generators/publish.ts",
+    "publish-sandboxes": "ts-node --swc ../scripts/sandbox/publish.ts",
     "publish:debug": "npm run publish:latest -- --npm-tag=debug --no-push",
     "publish:latest": "lerna publish --exact --concurrency 1 --force-publish",
     "publish:next": "npm run publish:latest -- --npm-tag=next",
