+++ generate-sandboxes.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 
-./scripts/node_modules/.bin/ts-node ./scripts/sandbox-generators/generate.ts
+./scripts/node_modules/.bin/ts-node ./scripts/sandbox/generate.ts
