+++ scripts/sandbox/generate.ts
