+++ scripts/sandbox/templates/root.ejs
