+++ scripts/sandbox/utils/template.ts
